+++ src/delegated-permissions.ts
@@ -2,15 +2,69 @@ import { createPermissionStatus } from "./delegated-permission-status.js";
 
 let canConstruct = false;
 
-export function createDelegatedPermissions({
-  delegates,
-}: {
+/**
+ * Parameters for creating a delegated Permissions API.
+ *
+ * @inline
+ * @see {@link createDelegatedPermissions} to create a delegated Permissions
+ *   API.
+ */
+export type DelegatedPermissionsParameters = {
+  /**
+   * The Permissions APIs to delegate to.
+   *
+   * The list must have at least one delegate, which will be selected initially.
+   * The list is static, and cannot be changed after the delegated API is
+   * created.
+   */
   delegates: globalThis.Permissions[];
-}): {
+};
+
+/**
+ * Create a Permissions API that delegates to other Permissions APIs.
+ *
+ * Delegated permissions can be used, for example, to create a Permissions API
+ * that "switches" between a fake Permissions API and a real Permissions API.
+ *
+ * When {@link @types/web!Permissions.query} is called on the delegated
+ * Permissions API, the resulting {@link PermissionStatus} will reflect the
+ * {@link PermissionState} of the permission in the selected delegate.
+ *
+ * Permissions API delegates can be selected dynamically at any time, and any
+ * {@link PermissionStatus} queries will immediately update to reflect the new
+ * delegate's {@link PermissionState} for the relevant permission, dispatching
+ * events as appropriate to any registered `change` listeners.
+ *
+ * @param params - The parameters for creating the delegated Permissions API.
+ *
+ * @returns The delegated Permissions API, and functions for managing the
+ *   selected delegate.
+ */
+export function createDelegatedPermissions(
+  params: DelegatedPermissionsParameters,
+): {
+  /**
+   * The delegated Permissions API.
+   */
   permissions: globalThis.Permissions;
-  selectDelegate: SelectDelegate;
-  isDelegateSelected: IsDelegateSelected;
+
+  /**
+   * Select a Permissions API delegate.
+   *
+   * @param delegate - The delegate to select.
+   */
+  selectDelegate: (delegate: globalThis.Permissions) => void;
+
+  /**
+   * Check if a Permissions API delegate is selected.
+   *
+   * @param delegate - The delegate to check.
+   *
+   * @returns `true` if the delegate is selected, `false` otherwise.
+   */
+  isSelectedDelegate: (delegate: globalThis.Permissions) => boolean;
 } {
+  const { delegates } = params;
   let [delegate] = delegates;
   if (!delegate) throw new TypeError("No delegates provided");
 
@@ -36,6 +90,10 @@ export function createDelegatedPermissions({
     }),
 
     selectDelegate(selectedDelegate) {
+      if (!delegates.includes(selectedDelegate)) {
+        throw new TypeError("Unknown delegate");
+      }
+
       delegate = selectedDelegate;
 
       for (const subscriber of subscribers) {
@@ -53,14 +111,11 @@ export function createDelegatedPermissions({
       }
     },
 
-    isDelegateSelected(query) {
+    isSelectedDelegate(query) {
       return query === delegate;
     },
   };
 }
-
-export type SelectDelegate = (delegate: globalThis.Permissions) => void;
-export type IsDelegateSelected = (delegate: globalThis.Permissions) => boolean;
 
 type PermissionsParameters = {
   delegates: globalThis.Permissions[];